+++ app/src/main/java/org/simple/clinic/user/LoggedInUserPayload.kt
@@ -30,5 +30,8 @@ data class LoggedInUserPayload(
     val createdAt: Instant,
 
     @Json(name = "updated_at")
-    val updatedAt: Instant
+    val updatedAt: Instant,
+
+    @Json(name = "teleconsult_phone_number")
+    val teleconsultPhoneNumber: String?
 )